+++ src/app/middlewares/errorHandler.ts
@@ -2,14 +2,14 @@
  * エラーハンドラーミドルウェア
  */
 import * as createDebug from 'debug';
-import { NextFunction, Request, Response } from 'express';
+import type { NextFunction, Request, Response } from 'express';
 import { INTERNAL_SERVER_ERROR } from 'http-status';
 
 import { APIError } from '../error/api';
 
 const debug = createDebug('surfrock-reserve-api:middlewares');
 
-export default (err: any, __: Request, res: Response, next: NextFunction) => {
+export async function errorHandler(err: any, __: Request, res: Response, next: NextFunction) {
     debug(err);
     if (res.headersSent) {
         next(err);
@@ -32,4 +32,4 @@ export default (err: any, __: Request, res: Response, next: NextFunction) => {
         .json({
             error: apiError.toObject()
         });
-};
+}