+++ src/app/middlewares/permitScopes.ts
@@ -2,7 +2,7 @@
  * スコープ許可ミドルウェア
  */
 import * as createDebug from 'debug';
-import { NextFunction, Request, Response } from 'express';
+import type { NextFunction, Request, Response } from 'express';
 import { UNAUTHORIZED } from 'http-status';
 
 import { } from '../../@types/index';
@@ -15,7 +15,7 @@ const debug = createDebug('surfrock-reserve-api:middlewares');
  */
 type IScope = string;
 
-export default (permittedScopes: IScope[]) => {
+export function permitScopes(permittedScopes: IScope[]) {
     return (req: Request, _: Response, next: NextFunction) => {
         if (process.env.RESOURCE_SERVER_IDENTIFIER === undefined) {
             next(new Error('RESOURCE_SERVER_IDENTIFIER undefined'));
@@ -48,7 +48,7 @@ export default (permittedScopes: IScope[]) => {
             next(error);
         }
     };
-};
+}
 
 /**
  * 所有スコープが許可されたスコープかどうか